fix(backend): fix operations and activities statistics
+ upgrade Elasticsearch to 6.6.0

Related with OPE-393

+++ backend/src/test/java/org/camunda/operate/util/TestUtil.java
@@ -88,7 +88,7 @@ public abstract class TestUtil {
     if (incidentKey != null) {
       activityInstanceForListViewEntity.setIncidentKey(incidentKey);
     } else {
-      activityInstanceForListViewEntity.setIncidentKey(random.nextLong());
+      activityInstanceForListViewEntity.setIncidentKey((long)random.nextInt());
     }
     if (errorMsg != null) {
       activityInstanceForListViewEntity.setErrorMessage(errorMsg);
